successfully added spells to backend via SRD API handler. began work on displaying spells as context vars on the spell sheet. have verified that it is doable (phew!) and will finish the rest tomorrow.
deliberation is required on how the rest of the spell data itself is going to be handled. perhaps a compendium page could be added? and perhaps it would be best to allow spells to just be stored as plaintext rather than being connected explicitly to models. again, though, this requires further deliberation. nightie-night.